add_angles_column.py

import asyncio
import os
from dotenv import load_dotenv
from supabase import create_client, Client

load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_SECRET_KEY")

if not SUPABASE_URL or not SUPABASE_KEY:
    print("Error: Supabase credentials not found in environment variables.")
    exit(1)

supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)

def add_angles_column():
    print("Attempting to add 'generated_angles' column to 'viral_videos' table...")
    
    # We can't execute raw SQL easily with the python client unless we use a function or just hope the user runs it.
    # However, sometimes we can use the `postgres` interface or just use a workaround.
    # Actually, the python client doesn't support raw SQL query execution directly unless enabled.
    # But we can try to use the rpc call if there is a 'exec_sql' function, which there isn't by default.
    
    # Alternative: check if column exists, if not, print instructions? 
    # Or creating a migration file is better documentation. 
    # But I need to EXECUTE it.
    
    # Let's try to use the `storage` or standard postgrest to see if we can do ddl? No.
    
    # WAIT. I can't easily alter table from the python client without a specific RPC function.
    # I should check if the user has a way to run SQL.
    # OR, I can use the existing `generated_scripts` table to store INTERMEDIATE results? No...
    
    # workaround: I will assume I CANNOT modify the schema easily from here without user intervention.
    # I will modify the flow to NOT depend on a new column if possible.
    # BUT, to implement the flow the USER wants (redirect to new page), I MUST persist data.
    
    # Alternative 2: Pass the ENTIRE angles JSON array in the URL or LocalStorage?
    # URL length limits might be hit.
    # LocalStorage is plausible.
    
    # Let's try to stick to server-side.
    # What if I store the angles in `generated_scripts` but mark it as "DRAFT" or "PENDING"?
    # The `generated_scripts` table requires `script`, `titles` which are NOT NULL. So I can't insert a partial record.
    
    # Okay, I will try to run the SQL via a specialized RPC if I can find one, OR I will just instruct the user.
    # OR... I will try to use the `pg` library or `psycopg2` if available? 
    # requirements.txt check.
    pass

if __name__ == "__main__":
    print("Checking dependencies...")
