requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.17
aiofiles==24.1.0

# AI Services
google-api-python-client==2.154.0
google-genai
anthropic==0.39.0

# Research & Scraping APIs
exa-py==1.0.10
firecrawl-py
perplexityai  # Perplexity AI SDK
openai==1.57.4  # Fallback for Perplexity (OpenAI-compatible)

# Other Dependencies
apify-client==2.3.0
pandas==2.2.3
requests==2.32.3
pillow==11.0.0
isodate==0.7.2
python-dotenv==1.0.1

# Additional Utilities
pydantic==2.10.0
pydantic-settings==2.6.0

# Database & Backend Services
supabase==2.10.0
stripe==11.2.0

# Authentication
authlib==1.3.2
httpx==0.27.2
itsdangerous==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==6.0.0